Fix wrong size of enum value (int instead of long)
Reviewed-on: http://gerrit.aldebaran.lan/28659
Tested-by: gerrit
Reviewed-on: http://gerrit.aldebaran.lan/28661
Tested-by: hcuche <hcuche@aldebaran-robotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/28860
Tested-by: llec <llec@aldebaran-robotics.com>

+++ qitype/typeinterface.hpp
@@ -207,9 +207,12 @@ namespace qi{
    dynamic_cast< ::qi::TypeOfTemplate<templateName>*>(typeInst)
 
 
+  /**
+   * TODO: Find the right size of enum values
+   */
   #define QI_TYPE_ENUM_REGISTER(Enum)                                \
     namespace qi {                                                   \
-      template<> class TypeImpl<Enum>: public IntTypeInterfaceImpl<long> {};  \
+      template<> class TypeImpl<Enum>: public IntTypeInterfaceImpl<int> {};  \
     }
 
 #define QI_TYPE_STRUCT_DECLARE(name)                                      \
